move libmythbase/ffmpeg-mmx.h to libmythtv/visualisations/goom/mmx.h
replacing the current mmx.h, whose extra values are unused
