+++ mythtv/libs/libmythbase/libmythbase.pro
@@ -30,7 +30,6 @@ HEADERS += mythscheduler.h filesysteminfo.h hardwareprofile.h serverpool.h
 HEADERS += mythbinaryplist.h signalhandling.h mythtimezone.h mythdate.h
 HEADERS += mythplugin.h mythpluginapi.h housekeeper.h
 HEADERS += mythpluginexport.h
-HEADERS += ffmpeg-mmx.h
 HEADERS += mythrandom.h
 HEADERS += stringutil.h
 HEADERS += mythsystemlegacy.h mythtypes.h
@@ -192,7 +191,7 @@ inc.files += mythcdrom.h autodeletedeque.h dbutil.h mythdeque.h
 inc.files += referencecounter.h referencecounterlist.h mythcommandlineparser.h
 inc.files += mthread.h mthreadpool.h mythchrono.h
 inc.files += filesysteminfo.h hardwareprofile.h bonjourregister.h serverpool.h
-inc.files += plist.h signalhandling.h ffmpeg-mmx.h mythdate.h
+inc.files += plist.h signalhandling.h mythdate.h
 inc.files += mythplugin.h mythpluginapi.h
 inc.files += mythpluginexport.h
 inc.files += remotefile.h mythsystemlegacy.h mythtypes.h
